move parallex from adapter to RecyclerView with the dismiss bug
