+++ UltimateRecyclerView/ultimaterecyclerview/src/main/java/com/marshalchen/ultimaterecyclerview/UltimateViewAdapter.java
@@ -20,15 +20,11 @@ import java.util.List;
 public abstract class UltimateViewAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
 
     private View customLoadMoreView = null;
-//    private View customHeaderView = null;
-//
-//    public void setCustomHeaderView(View customHeaderView) {
-//        this.customHeaderView = customHeaderView;
-//    }
 
     @Override
     public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-        if (viewType == 1) {
+
+        if (viewType == VIEW_TYPES.FOOTER) {
             if (customLoadMoreView == null) {
                 View v = LayoutInflater.from(parent.getContext())
                         .inflate(R.layout.bottom_progressbar, parent, false);
@@ -36,8 +32,9 @@ public abstract class UltimateViewAdapter extends RecyclerView.Adapter<RecyclerV
             } else {
                 return new ProgressBarViewHolder(customLoadMoreView);
             }
-        } else if (viewType == 2) {
-            if (mHeader != null) return new ProgressBarViewHolder(mHeader);
+        } else if (viewType == VIEW_TYPES.HEADER) {
+            if (UltimateRecyclerView.mHeader != null)
+                return new ProgressBarViewHolder(UltimateRecyclerView.mHeader);
         }
 
 
@@ -55,10 +52,11 @@ public abstract class UltimateViewAdapter extends RecyclerView.Adapter<RecyclerV
     @Override
     public int getItemViewType(int position) {
         if (position == getItemCount() - 1) {
-            return 1;
-        } else if (position == 0) return 2;
-        else
-            return super.getItemViewType(position);
+            return VIEW_TYPES.FOOTER;
+        } else if (position == 0) {
+            return VIEW_TYPES.HEADER;
+        } else
+            return VIEW_TYPES.NORMAL;
     }
 
 
@@ -108,7 +106,7 @@ public abstract class UltimateViewAdapter extends RecyclerView.Adapter<RecyclerV
     public <T> void insert(List<T> list, T object, int position) {
         list.add(position, object);
         notifyItemInserted(position);
-        notifyItemChanged(position + 1);
+        //  notifyItemChanged(position + 1);
     }
 
     public void remove(List<?> list, int position) {
@@ -122,82 +120,89 @@ public abstract class UltimateViewAdapter extends RecyclerView.Adapter<RecyclerV
         notifyItemRangeRemoved(0, size);
     }
 
-    private CustomRelativeWrapper mHeader;
-    private int mTotalYScrolled;
-    private final float SCROLL_MULTIPLIER = 0.5f;
-    private OnParallaxScroll mParallaxScroll;
-    public void setParallaxHeader(View header, RecyclerView view) {
-        // mRecyclerView = view;
-        mHeader = new CustomRelativeWrapper(header.getContext());
-        mHeader.setLayoutParams(new ViewGroup.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
-        mHeader.addView(header, new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
-        view.setOnScrollListener(new RecyclerView.OnScrollListener() {
-            @Override
-            public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
-                super.onScrolled(recyclerView, dx, dy);
-                if (mHeader != null) {
-                    mTotalYScrolled += dy;
-                    translateHeader(mTotalYScrolled);
-                }
-            }
-        });
-    }
+//    private UltimateRecyclerView.CustomRelativeWrapper mHeader;
 
-    public void setOnParallaxScroll(OnParallaxScroll parallaxScroll) {
-        mParallaxScroll = parallaxScroll;
-        mParallaxScroll.onParallaxScroll(0, 0, mHeader);
-    }
-    public void translateHeader(float of) {
-        float ofCalculated = of * SCROLL_MULTIPLIER;
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
-            mHeader.setTranslationY(ofCalculated);
-        } else {
-            TranslateAnimation anim = new TranslateAnimation(0, 0, ofCalculated, ofCalculated);
-            anim.setFillAfter(true);
-            anim.setDuration(0);
-            mHeader.startAnimation(anim);
-        }
-        mHeader.setClipY(Math.round(ofCalculated));
-        if (mParallaxScroll != null) {
-            float left = Math.min(1, ((ofCalculated) / (mHeader.getHeight() * SCROLL_MULTIPLIER)));
-            mParallaxScroll.onParallaxScroll(left, of, mHeader);
-        }
-    }
-    public interface OnParallaxScroll {
-        /**
-         * Event triggered when the parallax is being scrolled.
-         *
-         * @param percentage
-         * @param offset
-         * @param parallax
-         */
-        void onParallaxScroll(float percentage, float offset, View parallax);
-    }
-    static class CustomRelativeWrapper extends RelativeLayout {
+    //    private int mTotalYScrolled;
+////    private final float SCROLL_MULTIPLIER = 0.5f;
+////    private OnParallaxScroll mParallaxScroll;
+////
+//    public void setParallaxHeader(View header, RecyclerView view) {
+//        // mRecyclerView = view;
+//        mHeader = new UltimateRecyclerView.CustomRelativeWrapper(header.getContext());
+//        mHeader.setLayoutParams(new ViewGroup.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
+//        mHeader.addView(header, new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
+//        view.setOnScrollListener(new RecyclerView.OnScrollListener() {
+//            @Override
+//            public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+//                super.onScrolled(recyclerView, dx, dy);
+//                if (mHeader != null) {
+//                    Logs.d("mHeader height---"+mHeader.getHeight()+"  "+mHeader.getWidth());
+//
+//                    UltimateRecyclerView.mTotalYScrolled += dy;
+//                    translateHeader(UltimateRecyclerView.mTotalYScrolled);
+//                }
+//            }
+//        });
+//    }
 
-        private int mOffset;
-
-        public CustomRelativeWrapper(Context context) {
-            super(context);
-        }
-
-        @Override
-        protected void dispatchDraw(Canvas canvas) {
-            canvas.clipRect(new Rect(getLeft(), getTop(), getRight(), getBottom() + mOffset));
-            super.dispatchDraw(canvas);
-        }
-
-        public void setClipY(int offset) {
-            mOffset = offset;
-            invalidate();
-        }
-    }
+    //
+//    public void setOnParallaxScroll(OnParallaxScroll parallaxScroll) {
+//        mParallaxScroll = parallaxScroll;
+//        mParallaxScroll.onParallaxScroll(0, 0, mHeader);
+//    }
+//
+//    public void translateHeader(float of) {
+//        float ofCalculated = of * 0.5f;
+//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
+//            mHeader.setTranslationY(ofCalculated);
+//        } else {
+//            TranslateAnimation anim = new TranslateAnimation(0, 0, ofCalculated, ofCalculated);
+//            anim.setFillAfter(true);
+//            anim.setDuration(0);
+//            mHeader.startAnimation(anim);
+//        }
+//        mHeader.setClipY(Math.round(ofCalculated));
+////        if (mParallaxScroll != null) {
+////            float left = Math.min(1, ((ofCalculated) / (mHeader.getHeight() * SCROLL_MULTIPLIER)));
+////            mParallaxScroll.onParallaxScroll(left, of, mHeader);
+////        }
+//    }
+//
+//    public interface OnParallaxScroll {
+//        void onParallaxScroll(float percentage, float offset, View parallax);
+//    }
+//
+//    static class CustomRelativeWrapper extends RelativeLayout {
+//
+//        private int mOffset;
+//
+//        public CustomRelativeWrapper(Context context) {
+//            super(context);
+//        }
+//
+//        @Override
+//        protected void dispatchDraw(Canvas canvas) {
+//            canvas.clipRect(new Rect(getLeft(), getTop(), getRight(), getBottom() + mOffset));
+//            super.dispatchDraw(canvas);
+//        }
+//
+//        public void setClipY(int offset) {
+//            mOffset = offset;
+//            invalidate();
+//        }
+//    }
 
     class ProgressBarViewHolder extends RecyclerView.ViewHolder {
         public ProgressBarViewHolder(View itemView) {
             super(itemView);
         }
 
+    }
+
+    private class VIEW_TYPES {
+        public static final int NORMAL = 0;
+        public static final int HEADER = 1;
+        public static final int FOOTER = 2;
     }
